Split the REST API specific filters to a new file.
